requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncpg==0.29.0
psutil==5.9.6
